% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/destructure.R
\name{destructure}
\alias{destructure}
\title{Destructure an Object}
\usage{
destructure(object)
}
\description{
`destructure` is primarily used to help define how an
object should be prepared, or transformed, prior to a
destructuring assignment. This can be relevant for
objects which have unique subsetting semantics -- for
example, [numeric_version] objects.
}
\details{
Packages which would like to define special destructring
semantics for certain object classes can implement
methods for this class.
}
